# feat: avoid some copies in torch formatter