--- a/src/datasets/formatting/torch_formatter.py
+++ b/src/datasets/formatting/torch_formatter.py
@@ -130,22 +130,27 @@ class TorchFormatter(TensorFormatter[Mapping, "torch.Tensor", Mapping]):
                         return torch.from_numpy(value)
                     else:
                         # Fallback to safe conversion via Python ints
-                        return torch.tensor(value, dtype=target_dtype, **self.torch_tensor_kwargs)
+                        kwargs = self.torch_tensor_kwargs.copy()
+                        kwargs.setdefault('dtype', target_dtype)
+                        return torch.tensor(value, **kwargs)
                 else:
                     # Use zero-copy conversion for compatible integer types
                     if value.dtype != np.int64:
                         # Need dtype conversion, use as_tensor for efficiency
-                        return torch.as_tensor(value, dtype=target_dtype, **self.torch_tensor_kwargs)
+                        kwargs = self.torch_tensor_kwargs.copy()
+                        kwargs.setdefault('dtype', target_dtype)
+                        return torch.as_tensor(value, **kwargs)
                     else:
                         # Perfect match, zero-copy conversion
                         return torch.from_numpy(value)
 
             # Handle floating point types
             elif np.issubdtype(value.dtype, np.floating):
-                target_dtype = torch.float32
                 if value.dtype != np.float32:
                     # Need dtype conversion
-                    return torch.as_tensor(value, dtype=target_dtype, **self.torch_tensor_kwargs)
+                    kwargs = self.torch_tensor_kwargs.copy()
+                    kwargs.setdefault('dtype', torch.float32)
+                    return torch.as_tensor(value, **kwargs)
                 else:
                     # Zero-copy conversion
                     return torch.from_numpy(value)
@@ -157,9 +162,13 @@ class TorchFormatter(TensorFormatter[Mapping, "torch.Tensor", Mapping]):
         elif isinstance(value, np.number):
             if np.issubdtype(value.dtype, np.integer):
                 # Use torch.as_tensor for scalar conversion with dtype control
-                return torch.as_tensor(value, dtype=torch.int64, **self.torch_tensor_kwargs)
+                kwargs = self.torch_tensor_kwargs.copy()
+                kwargs.setdefault('dtype', torch.int64)
+                return torch.as_tensor(value, **kwargs)
             elif np.issubdtype(value.dtype, np.floating):
-                return torch.as_tensor(value, dtype=torch.float32, **self.torch_tensor_kwargs)
+                kwargs = self.torch_tensor_kwargs.copy()
+                kwargs.setdefault('dtype', torch.float32)
+                return torch.as_tensor(value, **kwargs)
             else:
                 return torch.as_tensor(value, **self.torch_tensor_kwargs)